fix: remove the explicit cast of snprintf

+++ src/nvim/tui/input.c
@@ -212,7 +212,7 @@ static void forward_simple_utf8(TermInput *input, TermKeyKey *key)
 
   while (*ptr) {
     if (*ptr == '<') {
-      int n = (size_t)snprintf(buf + len, sizeof(buf) - len, "<lt>");
+      int n = snprintf(buf + len, sizeof(buf) - len, "<lt>");
       assert(n >= 0 && n < sizeof(buf) - len);
       len += n;
     } else {
